Fix multu

+++ src/cpu.cpp
@@ -197,7 +197,7 @@ void CPU::execute_r_type(R_Instruction inst) {
             break;
         }
         case OpFunction::MULTU: {
-            U_Product product = static_cast<U_Product>(get_register(inst.src1)) * static_cast<U_Product>(get_register(inst.src2));
+            U_Product product = static_cast<U_Product>(static_cast<uint32_t>(get_register(inst.src1))) * static_cast<U_Product>(static_cast<uint32_t>(get_register(inst.src2)));
             LO = product & 0xFFFFFFFF;
             HI = (product >> 32) & 0xFFFFFFFF;
             advance_pc(4);
